Refactor import paths to use updated module structure for MQTT processing and correlation
